refactor(1013): remove ARPCommandManager facade (SC-001, position 42)
Extracts ARPCommandManager (289 LOC, 22 static methods) from MistHelper.py
monolith to src/device/arp_command_manager.py using the SC-001 fresh-extraction
facade pattern. Lazy `mh = importlib.import_module("MistHelper")` fetches
resolve live globals (apisession, PromptClientUtils, FilePathUtils) inside the
helpers that need them. Direct import for WebSocketStreamTarget dataclass
(already extracted).

- MistHelper.py -293 LOC (13218 -> 12925); re-export alias added alphabetically
  in the src.device.* block; class body replaced with single-line breadcrumb.
- Fail-fast install guards for `import websocket` and `import requests` annotated
  with `# noqa: F401` (both used exclusively by arp_command_manager after
  extraction, but preserved as module-load-time dependency checks).
- WebSocketStreamTarget import converted to re-export alias (# pylint: disable
  unused-import + # noqa: F401) to preserve MistHelper.WebSocketStreamTarget
  consumer surface.
- pyproject.toml: adds src/device/arp_command_manager.py to coverage omit.
- Local gate: ruff/black/py_compile clean; --test baseline 66/0/131 (unchanged).
